+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Edit the section for your next version **before** pushing the release tag like v
 
 ### Changes
   <!-- INSERT BULLETS UNDER THIS LINE -->
+- FEAT: Build and test windows-arm-desktop natively on windows-11-arm
 - FIX: make sure arm for windows is built
 - FEAT: Added Windws arm build
 - CHORE: Extract MSIX packaging into shared script, build it in CI